server/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
black==24.2.0
blinker==1.6.2
bs4==0.0.2
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dataclasses==0.6
dataclasses-serialization==1.3.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
flask-marshmallow==1.2.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-Session==0.5.0
Flask-Sessions==0.1.5
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
greenlet==2.0.2
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
loguru==0.7.0
lxml==5.1.0
Mako==1.3.2
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
more-properties==1.1.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pipdeptree==2.13.0
platformdirs==4.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
pyuwsgi==2.0.21
PyYAML==6.0.1
redis==5.0.0
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.19
toolz==0.12.0
toposort==1.10
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
